writing-practice/requirements.txt
streamlit==1.32.0
streamlit-drawable-canvas==0.9.3
manga-ocr==0.1.10
torch==2.2.0
pillow==10.2.0
requests==2.31.0